Read variables from `oceancolor` L3 bin files.

Compound vars we would usually want are "BinList" and `c("Rrs_443", "Rrs_488", "Rrs_555", "Rrs_547")` since these are used for calculating chlorophyll-a. This function is not very general yet, you'll always get the BinList.

read_L3_file(file, compound_vars = NULL)

Arguments

file
file name

compound_vars
the variables to read

Value

a tibble of the compound variables and bin details

Details

This function is specific to files in this form, for any of the platforms (MODISA here): https://oceandata.sci.gsfc.nasa.gov/MODIS-Aqua/L3BIN

Examples

read_L3_file(system.file("extdata/ocfiles/S2008001.L3b_DAY_RRS.nc", package = "roc"))

#> # A tibble: 2 x 26
#>   bin_num  nobs nscenes weights  time_rec BinList_bin_num BinList_nobs
#>     <int> <int>   <int>   <dbl>     <dbl>           <int>        <int>
#> 1   72251     1       1       1 473283776           72251            1
#> 2   89250     1       1       1 473295680           89250            1
#> # ... with 19 more variables: BinList_nscenes <int>, BinList_weights <dbl>,
#> #   BinList_time_rec <dbl>, Rrs_412_sum <dbl>, Rrs_412_sum_squared <dbl>,
#> #   Rrs_443_sum <dbl>, Rrs_443_sum_squared <dbl>, Rrs_490_sum <dbl>,
#> #   Rrs_490_sum_squared <dbl>, Rrs_510_sum <dbl>, Rrs_510_sum_squared <dbl>,
#> #   Rrs_555_sum <dbl>, Rrs_555_sum_squared <dbl>, Rrs_670_sum <dbl>,
#> #   Rrs_670_sum_squared <dbl>, angstrom_sum <dbl>, angstrom_sum_squared <dbl>,
#> #   aot_865_sum <dbl>, aot_865_sum_squared <dbl>